+++ src/auth/auth-wall.test.js
@@ -1,3 +1,4 @@
+import { NoticeBox } from '@dhis2/ui'
 import { shallow } from 'enzyme'
 import React from 'react'
 import { ErrorMessage } from '../shared/index.js'
@@ -14,15 +15,24 @@ afterEach(() => {
 
 describe('<AuthWall>', () => {
     it('shows a noticebox for unauthorized users', () => {
-        useIsAuthorized.mockImplementation(() => false)
+        useIsAuthorized.mockImplementation(() => ({
+            hasAppAccess: false,
+            hasApprovalAuthorities: false,
+        }))
 
         const wrapper = shallow(<AuthWall>Child</AuthWall>)
 
         expect(wrapper.find(ErrorMessage)).toHaveLength(1)
+        expect(wrapper.prop('children')).toBe(
+            "You don't have access to the Data Approval App. Contact a system administrator to request access."
+        )
     })
 
     it('renders the children for authorised users', () => {
-        useIsAuthorized.mockImplementation(() => true)
+        useIsAuthorized.mockImplementation(() => ({
+            hasAppAccess: true,
+            hasApprovalAuthorities: true,
+        }))
 
         const wrapper = shallow(<AuthWall>Child</AuthWall>)