+++ src/bottom-bar/bottom-bar.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { ApprovalStatusTag } from '../shared/index.js'
+import { useIsAuthorized } from '../auth/use-is-authorized.js'
+import { ApprovalStatusTag, APPROVAL_STATUSES } from '../shared/index.js'
 import { useWorkflowContext } from '../workflow-context/index.js'
 import { AcceptButton } from './accept-button/index.js'
 import { ApproveButton } from './approve-button/index.js'
@@ -8,18 +9,33 @@ import styles from './bottom-bar.module.css'
 import { UnacceptButton } from './unaccept-button/index.js'
 import { UnapproveButton } from './unapprove-button/index.js'
 
+const approvedStatuses = new Set([
+    APPROVAL_STATUSES.APPROVED_HERE,
+    APPROVAL_STATUSES.APPROVED_ABOVE,
+    APPROVAL_STATUSES.ACCEPTED_HERE,
+])
+
 const BottomBar = () => {
     const { allowedActions, approvalStatus, approvedBy, approvedAt } =
         useWorkflowContext()
+    const { hasApprovalAuthorities } = useIsAuthorized()
     const { mayAccept, mayApprove, mayUnaccept, mayUnapprove } = allowedActions
-    const disableApproveBtn = !mayApprove && !mayUnapprove
+    const disableApproveBtn =
+        /* We want to signal that the user can't approve or unapprove anything 
+           by showing a disabled button rather than an empty space */
+        (!mayApprove && !mayUnapprove) ||
+        (mayApprove && approvedStatuses.has(approvalStatus))
 
     return (
         <>
             <div className={styles.bottomBar} data-test="bottom-bar">
                 <BottomBarItem>
                     <ApprovalStatusTag
-                        approvalStatus={approvalStatus}
+                        approvalStatus={
+                            hasApprovalAuthorities
+                                ? approvalStatus
+                                : APPROVAL_STATUSES.UNAUTHORIZED
+                        }
                         approvedBy={approvedBy}
                         approvedAt={approvedAt}
                     />