+++ src/shared/approval-status/get-approval-status.test.js
@@ -1,6 +1,9 @@
 import { IconBlock16, IconError16 } from '@dhis2/ui'
 import moment from 'moment'
-import { getApprovalStatusDisplayData } from './get-approval-status.js'
+import {
+    getApprovalStatusDisplayData,
+    APPROVAL_STATUSES,
+} from './get-approval-status.js'
 import { Approved, Ready, Waiting } from './icons.js'
 
 jest.mock('moment', () => {
@@ -10,7 +13,9 @@ jest.mock('moment', () => {
 describe('getApprovalStatusDisplayData', () => {
     it('returns the correct display data for approval status "UNAPPROVED_READY"', () => {
         expect(
-            getApprovalStatusDisplayData({ approvalStatus: 'UNAPPROVED_READY' })
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.UNAPPROVED_READY,
+            })
         ).toEqual({
             displayName: 'Ready for approval',
             icon: Ready,
@@ -19,17 +24,20 @@ describe('getApprovalStatusDisplayData', () => {
     })
     it('returns the correct display data for approval status "ACCEPTED_HERE"', () => {
         expect(
-            getApprovalStatusDisplayData({ approvalStatus: 'ACCEPTED_HERE' })
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.ACCEPTED_HERE,
+                approvedBy: 'Hendrik',
+            })
         ).toEqual({
-            displayName: 'Ready for approval — Accepted',
-            icon: Ready,
-            type: 'neutral',
+            displayName: 'Approval by Hendrik accepted 2 years ago',
+            icon: Approved,
+            type: 'positive',
         })
     })
     it('returns the correct display data for approval status "UNAPPROVED_WAITING"', () => {
         expect(
             getApprovalStatusDisplayData({
-                approvalStatus: 'UNAPPROVED_WAITING',
+                approvalStatus: APPROVAL_STATUSES.UNAPPROVED_WAITING,
             })
         ).toEqual({
             displayName: 'Waiting for lower level approval',
@@ -39,66 +47,56 @@ describe('getApprovalStatusDisplayData', () => {
     })
     it('returns the correct display data for approval status "UNAPPROVED_ABOVE"', () => {
         expect(
-            getApprovalStatusDisplayData({ approvalStatus: 'UNAPPROVED_ABOVE' })
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.UNAPPROVED_ABOVE,
+            })
         ).toEqual({
             displayName: 'Waiting for higher level approval',
             icon: Waiting,
             type: 'default',
         })
     })
-    describe('approved approval statuses "APPROVED_HERE" and "APPROVED_ABOVE"', () => {
-        for (const approvalStatus of ['APPROVED_HERE', 'APPROVED_ABOVE']) {
-            it(`returns the correct diplay data for ${approvalStatus} when only approvalStatus is supplied`, () => {
-                expect(
-                    getApprovalStatusDisplayData({ approvalStatus })
-                ).toEqual({
-                    displayName: 'Approved',
-                    icon: Approved,
-                    type: 'positive',
-                })
+    it('returns the correct diplay data for "APPROVED_HERE"', () => {
+        expect(
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.APPROVED_HERE,
+                approvedBy: 'Hendrik',
+                // The actual value for this field is irrelevant due to the moment mock
+                approvalDateTime: 'Not empty',
             })
-            it(`returns the correct diplay data for ${approvalStatus} when approvalStatus and approvedBy are passed`, () => {
-                expect(
-                    getApprovalStatusDisplayData({
-                        approvalStatus,
-                        approvedBy: 'Hendrik',
-                    })
-                ).toEqual({
-                    displayName: 'Approved by Hendrik',
-                    icon: Approved,
-                    type: 'positive',
-                })
+        ).toEqual({
+            displayName: 'Approved by Hendrik 2 years ago',
+            icon: Approved,
+            type: 'positive',
+        })
+    })
+    it('returns the correct diplay data for "APPROVED_ABOVE"', () => {
+        expect(
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.APPROVED_ABOVE,
             })
-            it(`returns the correct diplay data for ${approvalStatus} when approvalStatus and approvalDateTime are passed`, () => {
-                expect(
-                    getApprovalStatusDisplayData({
-                        approvalStatus,
-                        approvalDateTime: '2020-08-24T18:55:03.165Z',
-                    })
-                ).toEqual({
-                    displayName: 'Approved 2 years ago',
-                    icon: Approved,
-                    type: 'positive',
-                })
+        ).toEqual({
+            displayName: 'Approved at higher level',
+            icon: Approved,
+            type: 'positive',
+        })
+    })
+    it('returns the correct display data for approval status "UNAUTHORIZED"', () => {
+        expect(
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.UNAUTHORIZED,
             })
-            it(`returns the correct diplay data for ${approvalStatus} when only all options are passed`, () => {
-                expect(
-                    getApprovalStatusDisplayData({
-                        approvalStatus,
-                        approvedBy: 'Hendrik',
-                        approvalDateTime: '2020-08-24T18:55:03.165Z',
-                    })
-                ).toEqual({
-                    displayName: 'Approved by Hendrik 2 years ago',
-                    icon: Approved,
-                    type: 'positive',
-                })
-            })
-        }
+        ).toEqual({
+            displayName: 'You do not have authority to approve data',
+            icon: IconBlock16,
+            type: 'negative',
+        })
     })
     it('returns the correct display data for approval status "UNAPPROVABLE"', () => {
         expect(
-            getApprovalStatusDisplayData({ approvalStatus: 'UNAPPROVABLE' })
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.UNAPPROVABLE,
+            })
         ).toEqual({
             displayName: 'Cannot be approved',
             icon: IconBlock16,
@@ -107,7 +105,9 @@ describe('getApprovalStatusDisplayData', () => {
     })
     it('returns the correct display data for approval status "ERROR"', () => {
         expect(
-            getApprovalStatusDisplayData({ approvalStatus: 'ERROR' })
+            getApprovalStatusDisplayData({
+                approvalStatus: APPROVAL_STATUSES.ERROR,
+            })
         ).toEqual({
             displayName: 'Could not retrieve approval status',
             icon: IconError16,